20230325 icons (#349)
* Add folder and search icons

* Archive meta mask component

* Spelling and corrections

* update flex
