+++ .vscode/settings.json
@@ -56,7 +56,8 @@
     "Sepolia",
     "permissionlessly",
     "permissionless",
-    "Gelato"
+    "Gelato",
+    "incentivizing"
   ],
   "editor.defaultFormatter": "esbenp.prettier-vscode",
   "editor.formatOnSave": true,
